音乐播放

+++ src/jvmMain/kotlin/base/player/IPlayerListener.kt
@@ -6,5 +6,5 @@ package base.player
  */
 interface IPlayerListener {
     fun onStatusChanged(status: PlayerStatus)
-    fun onProgress(totalDuring: Int, currentPosition: Int, percentage: Int)
+    fun onProgress(totalDuring: Int, currentPosition: Int, percentage: Float)
 }
